Cleanup `pc_repository` module
Instead of collating the packagecontrol registry on the client, we
actually do this on the server using a GitHub action.

As a consequence the `pc_repository` module can be simplified.  The
ready time for the dashboard goes down from 3s to <1s.  (However, the
3s were actually too fast, and Sublime Text stuttered.  So, in reality,
I slowed down the local crawler to took roughly 12-14s(!) to run.)